Add code for clicking operators, equal sign, tweak
Wrote the code that causes the first value that is entered to be saved from the variable a into the varible b, when one of the operator buttons are clicked. Then the second value is saved into the variable a. The display is also cleared, when clicking one of the operators. Also referenced the equal sign and made it so that clicking it logs the result in the console. The operation depends on the operator clicked. Still have to add code to display result in the display.

Also in the function operate I had to add return before the call of each function, because undefined was logged in the console instead of the result.