Add result to display, enable further calculation
Result is displayed on the display. Also edited the operators event listener code so that when the first result is displayed, the value is stored into the variable b. Then by clicking the operators and selecting the value for a, further calculation is possible. Before, when you wanted to calculate further, when you clicked and operator, the value of b became the same as the previous a value.