Switch a and b in subtraction and division
Previously, the subtract and divide functions subtracted b from a and divided a with b. Because the first value entered in the calculator was saved into the variable b, this needed to be changed, so that the first number you enter into the calculator is divided by the second and the second number you entered in the calculator was subtracted from the first.