Fix operators, equals and clear listeners
The change in code makes it so that the action of clicking an operator depends on if an operator button was already clicked or not. If it was not clicked it works as before, the value of variable b is set to the value of a and the value of a is reset to "". But if the operator is clicked for the second time, before the equals operator is clicked, the two values written so far will be operated on and will return a result. The variable b will save the value of the result, the variable a will be reset and ready for further calculation.

The equals button was changed to work similarly to the operator buttons, when clicked the second time.

The clear button was fixed, so it resets the calculator completely. Before, not all the variables a, b and operator were reset to "".