Fix appending to display issue when typing digits
When the first number was input and the operator clicked, inputting the second number would just append the digits to the first number. Now by introducing the shouldResetDisplay variable, the calculator knows when to clear the display and start the new number, aka after pressing the operator, and when to append digits to create a number.
But the problem still remains when trying to write a new number after getting a result. The digits just get appended to the result. Now to fix this