Stop appending to result when entering digits
When the result was displayed on display and new digits were entered, the digits were being appended to the result. shouldResetDisplay = true solved that, but then, when beginning a new calculation, the variable b became undefined and the calculator returned NaN to the new calculation. With the code added to the operator handler this was fixed and now immidiately after getting a result, a new calculation can be made without pressing the clear button.