Introduce variable lastAction
With this new variable, last action aka last button pressed is marked. So if the last button pressed is the equals button, then pressing the digit buttons will start a new number, new calculation from scratch. If the operator button is pressed after the equals button, chaining of calculations will be possible.